p2p/sentry/libsentry, node/direct: dedupe sentry stream adapters
Move the generic SentryStreamS/SentryStreamC/StreamReply adapters from
sentrymultiplexer.go into their own libsentry/stream.go, and use them from
SentryClientDirect.Messages/PeerEvents instead of the hand-monomorphized
SentryMessagesStreamS/C and SentryPeersStreamS/C copies (identical Send
with EvictOldestIfHalfFull, Err, Recv, RecvMsg with proto.Merge), which
are deleted. The exported copies were referenced only by their own tests.

The two node/direct eviction tests pinned one monomorphic copy each; with
a single generic implementation left, they collapse into one test moved to
libsentry/stream_test.go, keeping the stronger drop-oldest assertions.
