+++ p2p/sentry/libsentry/sentrymultiplexer.go
@@ -29,8 +29,6 @@ import (
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
-	"google.golang.org/protobuf/proto"
-	"google.golang.org/protobuf/reflect/protoreflect"
 	"google.golang.org/protobuf/types/known/emptypb"
 
 	"github.com/erigontech/erigon/common"
@@ -437,68 +435,6 @@ func (m *sentryMultiplexer) SendMessageToAll(ctx context.Context, in *sentryprot
 	}
 
 	return &sentryproto.SentPeers{Peers: allSentPeers}, nil
-}
-
-type StreamReply[T protoreflect.ProtoMessage] struct {
-	R   T
-	Err error
-}
-
-// SentryMessagesStreamS implements proto_sentry.Sentry_ReceiveMessagesServer
-type SentryStreamS[T protoreflect.ProtoMessage] struct {
-	Ch  chan StreamReply[T]
-	Ctx context.Context
-	grpc.ServerStream
-}
-
-func (s *SentryStreamS[T]) Send(m T) error {
-	s.Ch <- StreamReply[T]{R: m}
-	EvictOldestIfHalfFull(s.Ch)
-	return nil
-}
-
-func (s *SentryStreamS[T]) Context() context.Context { return s.Ctx }
-
-func (s *SentryStreamS[T]) Err(err error) {
-	if err == nil {
-		return
-	}
-	s.Ch <- StreamReply[T]{Err: err}
-}
-
-func (s *SentryStreamS[T]) Close() {
-	if s.Ch != nil {
-		ch := s.Ch
-		s.Ch = nil
-		close(ch)
-	}
-}
-
-type SentryStreamC[T protoreflect.ProtoMessage] struct {
-	Ch  chan StreamReply[T]
-	Ctx context.Context
-	grpc.ClientStream
-}
-
-func (c *SentryStreamC[T]) Recv() (T, error) {
-	m, ok := <-c.Ch
-	if !ok {
-		var t T
-		return t, io.EOF
-	}
-	return m.R, m.Err
-}
-
-func (c *SentryStreamC[T]) Context() context.Context { return c.Ctx }
-
-func (c *SentryStreamC[T]) RecvMsg(anyMessage any) error {
-	m, err := c.Recv()
-	if err != nil {
-		return err
-	}
-	outMessage := anyMessage.(T)
-	proto.Merge(outMessage, m)
-	return nil
 }
 
 func (m *sentryMultiplexer) Messages(ctx context.Context, in *sentryproto.MessagesRequest, opts ...grpc.CallOption) (sentryproto.Sentry_MessagesClient, error) {
